// src/emission/enum/sourcename.enum.ts
/**
 * Please keep only the completed emission source
 */
export enum sourceName {
  Electricity = 'electricity', 
  Generator = "generator", 
  FireExtinguisher='fire_extinguisher', 
  Refrigerant = 'refrigerant', 
  GasBiomass = 'gas_biomass',
  WeldingEs = 'welding_es', 
  Forklifts = 'forklifts', 
  Boilers = 'boiler', 
  Waste_water_treatment = 'waste_water_treatment', 
  Municipal_water = 'municipal_water', 
  waste_disposal = 'waste_disposal', 
  cooking_gas = 'cooking_gas',
  // Freight_transport = 'freight_transport',  


  freight_air = 'freight_air',
  freight_offroad = 'freight_offroad',
  freight_rail = 'freight_rail', 
  freight_road = 'freight_road', 
  freight_water = 'freight_water',

  passenger_air = 'passenger_air',//no equation
  passenger_offroad = 'passenger_offroad', 
  passenger_rail = 'passenger_rail', 
  passenger_road = 'passenger_road', 
  passenger_water = 'passenger_water', 

  offroad_machinery = 'offroad_machinery_offroad' ,

  t_n_d_loss = 't_n_d_loss',
  business_travel = 'business_travel', 


  Purchased_Goods_and_Services="purchased_goods_and_services",
  Capital_Goods = "capital_goods",
  Fuel_Energy_Related_Activities = "fuel_energy_related_activities",
  Upstream_Transportation_and_Distribution = "upstream_transportation_and_distribution",
  Waste_Generated_in_Operations = "waste_generated_in_operations",
  Net_Zero_Business_Travel = "net_zero_business_travel",
  Net_Zero_Employee_Commuting = "net_zero_employee_commuting",
  Upstream_Leased_Assets = "upstream_leased_assets",
  Downstream_Transportation_and_Distribution = "downstream_transportation_and_distribution",
  Processing_of_Sold_Products = "processing_of_sold_products",
  Use_of_Sold_Products = "use_of_sold_products",
  End_of_Life_Treatment_of_Sold_Products = "end_of_life_treatment_of_sold_products",
  Downstream_Leased_Assets = "downstream_leased_assets",
  Franchises = "franchises",
  Investments = "investments" 


}